organized similarity functions and added shared util methods
